Update CMSG_CAST_SPELL and SMSG_NEW_WORLD

+++ WowPacketParserModule.V7_0_3_22248/Parsers/SpellHandler.cs
@@ -132,7 +132,7 @@ namespace WowPacketParserModule.V7_0_3_22248.Parsers
 
             int mapID = -1;
             if (hasMapID)
-                mapID = (ushort)packet.ReadInt32("MapID", idx);
+                mapID = packet.ReadInt32("MapID", idx);
 
             if (Settings.UseDBC && dbdata.DstPosition != null && mapID != -1)
             {
